+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @aws-amplify/backend-cli
 
+## 1.6.0
+
+### Minor Changes
+
+- 8f59d16: integrate with aws cdk toolkit
+- 7b9a9d2: adding seed feature
+
+### Patch Changes
+
+- 7962bc5: Add notices
+- b7f6719: Cache dependencies when assessing notices predicates
+- b2f9042: fix: exclude logs streaming options from sandbox command help output
+- 2bd90e5: remove unnecessary toolkit dependencies
+- Updated dependencies [29a6167]
+- Updated dependencies [f9f31c7]
+- Updated dependencies [4c8829d]
+- Updated dependencies [f0d501e]
+- Updated dependencies [8f59d16]
+- Updated dependencies [f62a1fd]
+- Updated dependencies [0cc2de3]
+- Updated dependencies [b2f9042]
+- Updated dependencies [e4383ec]
+- Updated dependencies [9a00a6b]
+- Updated dependencies [1d2b36f]
+- Updated dependencies [1d2b36f]
+  - @aws-amplify/form-generator@1.1.0
+  - @aws-amplify/cli-core@2.0.0
+  - @aws-amplify/sandbox@2.0.0
+  - @aws-amplify/backend-deployer@2.0.0
+  - @aws-amplify/plugin-types@1.9.0
+  - @aws-amplify/client-config@1.6.0
+  - @aws-amplify/deployed-backend-client@1.6.0
+  - @aws-amplify/model-generator@1.1.0
+  - @aws-amplify/platform-core@1.7.0
+  - @aws-amplify/backend-secret@1.3.0
+  - @aws-amplify/backend-output-schemas@1.5.0
+  - @aws-amplify/schema-generator@1.3.0
+
 ## 1.5.0
 
 ### Minor Changes